pyproject.toml
[project]
name = "recommendations-places"
version = "0.1.0"
description = ""
authors = [
    {name = "ronihdzz",email = "ronaldo.runing_@hotmail.com"}
]
package-mode = false
requires-python = ">=3.12,<4.0"
dependencies = [
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "sentry-sdk (>=2.33.2,<3.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "openai (>=1.97.1,<2.0.0)",
    "qdrant-client (>=1.12.1,<2.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "tabulate (>=0.9.0,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"

